requirements.txt
openai>=1.40
pymupdf
pdfminer.six
pytesseract
Pillow
pandas
python-dotenv
tqdm
